moved and made a pynn 0.8 version
