+++ examples_7/.gitignore
