+++ examples_7/board_test.py
