+++ examples_7/external_device_examples/.gitignore
