+++ examples_7/external_device_examples/live_examples/spike_io.colour_map
