+++ examples_7/external_device_examples/live_examples/spike_io.py
@@ -1,18 +1,13 @@
 # imports of both spynnaker and external device plugin.
-import spynnaker.pyNN as Frontend
-import spynnaker_external_devices_plugin.pyNN as ExternalDevices
-
-#######################
-# import to allow prefix type for the prefix eieio protocol
-######################
-from spynnaker_external_devices_plugin.pyNN.connections\
-    .spynnaker_live_spikes_connection import SpynnakerLiveSpikesConnection
-
-# plotter in python
 import pylab
-import time
 import random
+import spynnaker7.pyNN as Frontend
+import time
 from threading import Condition
+
+import spynnaker_external_devices_plugin.pyNN as ExternalDevices
+from spynnaker_external_devices_plugin.pyNN.connections \
+    .spynnaker_live_spikes_connection import SpynnakerLiveSpikesConnection
 
 # boolean allowing users to use python or c vis
 using_c_vis = False
