+++ examples_7/external_device_examples/live_examples/synfire_if_curr_exp_live.py
