+++ examples_7/external_device_examples/retina_example.py
@@ -5,9 +5,9 @@ intermediate population
 try:
     import pyNN.spiNNaker as p
 except Exception as e:
-    import spynnaker.pyNN as p
-import spynnaker_external_devices_plugin.pyNN as q
+    import spynnaker7.pyNN as p
 import retina_lib as retina_lib
+import spynnaker_external_devices_plugin.pyNN as q
 
 connected_chip_details = {
     "spinnaker_link_id": 0,
