+++ examples_7/synfire_if_curr_exp.py
@@ -4,12 +4,13 @@ Synfirechain-like example
 try:
     import pyNN.spiNNaker as p
 except Exception as e:
-    import spynnaker.pyNN as p
+    import spynnaker7.pyNN as p
 import pylab
 
 p.setup(timestep=1.0, min_delay=1.0, max_delay=144.0)
 nNeurons = 200  # number of neurons in each population
-p.set_number_of_neurons_per_core("IF_cond_exp", nNeurons / 2)
+p.set_number_of_neurons_per_core("IF_curr_exp", nNeurons / 2)
+
 
 cell_params_lif = {'cm': 0.25,
                    'i_offset': 0.0,
@@ -19,15 +20,13 @@ cell_params_lif = {'cm': 0.25,
                    'tau_syn_I': 5.0,
                    'v_reset': -70.0,
                    'v_rest': -65.0,
-                   'v_thresh': -50.0,
-                   'e_rev_E': 0.,
-                   'e_rev_I': -80.
+                   'v_thresh': -50.0
                    }
 
 populations = list()
 projections = list()
 
-weight_to_spike = 0.035
+weight_to_spike = 2.0
 delay = 17
 
 loopConnections = list()
@@ -37,15 +36,15 @@ for i in range(0, nNeurons):
 
 injectionConnection = [(0, 0, weight_to_spike, 1)]
 spikeArray = {'spike_times': [[0]]}
-populations.append(p.Population(nNeurons, p.IF_cond_exp, cell_params_lif,
-                                label='pop_1'))
+populations.append(p.Population(nNeurons, p.IF_curr_exp, cell_params_lif,
+                   label='pop_1'))
 populations.append(p.Population(1, p.SpikeSourceArray, spikeArray,
-                                label='inputSpikes_1'))
+                   label='inputSpikes_1'))
 
 projections.append(p.Projection(populations[0], populations[0],
-                                p.FromListConnector(loopConnections)))
+                   p.FromListConnector(loopConnections)))
 projections.append(p.Projection(populations[1], populations[0],
-                                p.FromListConnector(injectionConnection)))
+                   p.FromListConnector(injectionConnection)))
 
 populations[0].record_v()
 populations[0].record_gsyn()
@@ -71,6 +70,8 @@ if spikes is not None:
     pylab.show()
 else:
     print "No spikes received"
+
+# Make some graphs
 
 if v is not None:
     ticks = len(v) / nNeurons
