+++ examples_7/synfire_if_curr_exp_large_array.py
@@ -4,7 +4,7 @@ Synfirechain-like example
 try:
     import pyNN.spiNNaker as p
 except Exception as e:
-    import spynnaker.pyNN as p
+    import spynnaker7.pyNN as p
 import pylab
 
 p.setup(timestep=1.0, min_delay=1.0, max_delay=144.0)
