+++ examples_8/synfire_if_curr_exp.py
@@ -1,16 +1,12 @@
 """
 Synfirechain-like example
 """
-try:
-    import pyNN.spiNNaker as p
-except Exception as e:
-    import spynnaker.pyNN as p
+import spynnaker8.pyNN as p
 import pylab
 
 p.setup(timestep=1.0, min_delay=1.0, max_delay=144.0)
 nNeurons = 200  # number of neurons in each population
 p.set_number_of_neurons_per_core("IF_curr_exp", nNeurons / 2)
-
 
 cell_params_lif = {'cm': 0.25,
                    'i_offset': 0.0,
@@ -31,24 +27,24 @@ delay = 17
 
 loopConnections = list()
 for i in range(0, nNeurons):
-    singleConnection = (i, ((i + 1) % nNeurons), weight_to_spike, delay)
+    singleConnection = (i, ((i + 1) % nNeurons))
     loopConnections.append(singleConnection)
 
-injectionConnection = [(0, 0, weight_to_spike, 1)]
+injectionConnection = [(0, 0)]
 spikeArray = {'spike_times': [[0]]}
-populations.append(p.Population(nNeurons, p.IF_curr_exp, cell_params_lif,
-                   label='pop_1'))
-populations.append(p.Population(1, p.SpikeSourceArray, spikeArray,
-                   label='inputSpikes_1'))
+populations.append(
+    p.Population(nNeurons, p.IF_curr_exp(**cell_params_lif), label='pop_1'))
+populations.append(
+    p.Population(1, p.SpikeSourceArray(**spikeArray), label='inputSpikes_1'))
 
-projections.append(p.Projection(populations[0], populations[0],
-                   p.FromListConnector(loopConnections)))
-projections.append(p.Projection(populations[1], populations[0],
-                   p.FromListConnector(injectionConnection)))
+projections.append(p.Projection(
+    populations[0], populations[0], p.FromListConnector(loopConnections),
+    p.StaticSynapse(weight=weight_to_spike, delay=delay)))
+projections.append(p.Projection(
+    populations[1], populations[0], p.FromListConnector(injectionConnection),
+    p.StaticSynapse(weight=weight_to_spike, delay=1)))
 
-populations[0].record_v()
-populations[0].record_gsyn()
-populations[0].record()
+populations[0].record(['v', 'gsyn_exc', 'gsyn_inh', 'spikes'])
 
 p.run(5000)
 
@@ -56,9 +52,10 @@ v = None
 gsyn = None
 spikes = None
 
-v = populations[0].get_v(compatible_output=True)
-gsyn = populations[0].get_gsyn(compatible_output=True)
-spikes = populations[0].getSpikes(compatible_output=True)
+v = populations[0].get('v', compatible_output=True)
+gsyn_exc = populations[0].get('gsyn_exc', compatible_output=True)
+gsyn_inh = populations[0].get('gsyn_inh', compatible_output=True)
+spikes = populations[0].get('spikes', compatible_output=True)
 
 if spikes is not None:
     print spikes
@@ -84,14 +81,25 @@ if v is not None:
         pylab.plot([i[2] for i in v_for_neuron])
     pylab.show()
 
-if gsyn is not None:
-    ticks = len(gsyn) / nNeurons
+if gsyn_exc is not None:
+    ticks = len(gsyn_exc) / nNeurons
     pylab.figure()
     pylab.xlabel('Time/ms')
-    pylab.ylabel('gsyn')
-    pylab.title('gsyn')
+    pylab.ylabel('gsyn exc')
+    pylab.title('gsyn exc')
     for pos in range(0, nNeurons, 20):
-        gsyn_for_neuron = gsyn[pos * ticks: (pos + 1) * ticks]
+        gsyn_for_neuron = gsyn_exc[pos * ticks: (pos + 1) * ticks]
+        pylab.plot([i[2] for i in gsyn_for_neuron])
+    pylab.show()
+
+if gsyn_inh is not None:
+    ticks = len(gsyn_inh) / nNeurons
+    pylab.figure()
+    pylab.xlabel('Time/ms')
+    pylab.ylabel('gsyn inh')
+    pylab.title('gsyn inh')
+    for pos in range(0, nNeurons, 20):
+        gsyn_for_neuron = gsyn_inh[pos * ticks: (pos + 1) * ticks]
         pylab.plot([i[2] for i in gsyn_for_neuron])
     pylab.show()
